feat: update date formatting; change month and weekday display to long format

+++ js/app.js
@@ -408,9 +408,9 @@ const App = (() => {
     const [year, month, day] = dateKey.split('-').map(Number);
     const date = new Date(year, month - 1, day);
     const formattedDate = date.toLocaleDateString('te-IN', {
-      month: 'short',
+      month: 'long',
       day: 'numeric',
-      weekday: 'short'
+      weekday: 'long'
     });
     const isSelected = dateKey === state.selectedDate;
     const newsCount = getNewsCountForDate(dateKey);
